Change org namespace for image on Docker Hub
The condition for pushing to Docker Hub was refined such that pushes are
only triggered for push events on the master branch of the drand/drand
repo. Using the "deploy" key instead of "after_success" made this simpler
since deployments are not executed for pull requests events.

The .travis.yml was also cleaned up a bit. The use of sudo was removed
as it does not have any effect anymore.

Fixes #201