feat: add failover client
`NewFailoverWatcher` creates a client whose Watch function will failover to `Get`-ing new randomness if it does not receive it after the passed grace period.

Note: with the failover watcher - if network conditions slow down the delivery of a round, the client will skip as many rounds as the duration of the network condition. i.e. the failover is currently only catering for the case where a pubsub message is not received, not when network issues occur.