fix: nil pointer when using failover without passing chainInfo (#412)
When `chainInfo` is not passed it's retrieved from `/info` and set on `cfg.chainInfo`. The `NewFailoverWatcher` constructor was being called outside of `makeClient` so although `makeClient` retrieved `chainInfo` and set it on `cfg.chainInfo`, it was not observed by `NewFailoverWatcher` because `makeClient` is passed a _copy_ of `clientConfig` not a pointer.

This PR moves the `NewFailoverWatcher` constructor into `makeClient`, as well as `NewCachingClient` and `newWatchAggregator`.