fix: cache for latest value (#443)
A `Get` with round = 0 implies we want the latest round, not the genesis round. This PR changes the HTTP client to request from `/public/latest` when round 0 is requested and `/public/{n}` otherwise. It also updates the caching client cache key so that a result is stored against it's round number as specified in the result, not against the round number requested (which could be 0!).