refactor: gossip use drand logger
This PR refactors the gossip client to use the drand logger instead of the ipfs logger. It changes some function args to accept a logger so that a custom logger can be passed to gossip client components. It also fixes some calls to the logger that had an uneven number of args.