feat: optimized watcher (#576)
This PR un-privileges the watcher client so that it's just another client passed to the optimizing client. It means that a client configured with a pubsub can also be configured with HTTP/gRPC and calls to `Watch` will race them all.

Currently a client `WithPubSub` will only watch for updates over pubsub, you need to configure the failover grace period to allow fetching over HTTP. This PR would allow us to remove the failover watcher client and just use the functionality in the optimizing client instead.