fix: retry auto watch (#702)
The optimizing client will retry failed clients but if _all_ of them fail and are put in backoff then the watch channel will be closed. If this happens then the auto watch will stop forever and there's no way to restart without creating a new client.

This PR changes the auto watch functionality so that if the auto watch channel gets closed, then it is re-opened after 30s. It adds `WithAutoWatchRetry(time.Duration)` option to allow people to customise or disable it.

Maybe this should be a separate client? It's not really tied to the aggregator anymore.