Expose own metrics in /metrics/peer/ (#754)
Rationaly here is that any metric scraper should support be configured in the
same way regardless of the peer. By skipping a handler for the current peer we
are creating an "exception" unique to every peer, which only makes things more
difficult to maintain on infra-side of things.

Also, chances are the index mapping is somehow borked, since in my tests it is
sometimes skipping the current node and sometimes a different node, so this might
fix that altogether.