dkgInfo set before echobroadcast attempted to connect isolated nodes
dkgInfo was being set on the beacon process after the echobroadcast.
This meant that isolated nodes were not actually informed of the DKG,
and would get excluded in future rounds.
The fix was to reorder some of the setting of dkg info (and locking) in
the dkg and reshare so it would be available for broadcast