Removed a slow and unnecessary DB length check in status command
Remote status calls print out some info about each node's view of the
world, including the current number of beacons in their database.
The remote status call requests a beacon DB length check but does not
actually use it in the corresponding CLI output.

I've replaced its value with one interpolated from the beacon count for backwards
compatibility as it's in the protobuf message, but it could probably be
removed